pyproject.toml
[tool.poetry]
name = "fast-api"
version = "0.1.0"
description = ""
authors = ["Mykyta-Harashchenko <garasenkonikita538@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
asyncpg = "^0.29.0"
uvicorn = "^0.30.5"
pydantic = {extras = ["email"], version = "^2.8.2"}
passlib = "^1.7.4"
fastapi-mail = "^1.4.1"
redis = "^5.0.8"
fastapi-limiter = "^0.1.6"
cloudinary = "^1.41.0"
jinja2 = "^3.1.4"
sphinx = "^8.0.2"


[tool.poetry.group.dev.dependencies]
sphinx = "^8.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
